chore: move tology related files under topology folder
