+++ .github/actions/deploy-components/action.yaml
@@ -37,6 +37,10 @@ inputs:
     description: 'Set slim controller container image version'
     required: false
     default: ''
+  slim-controller-chart-tag:
+    description: 'Set slim controller chart version (OCI only; ignored when slim-repo-path is set)'
+    required: false
+    default: ''
   deploy-directory:
     description: 'Deploy directory to a kind cluster'
     required: false
@@ -130,6 +134,7 @@ runs:
           SLIM_REPO_PATH='${{ inputs.slim-repo-path }}' \
           SLIM_IMAGE_TAG=${{ inputs.slim-image-tag }} \
           SLIM_CONTROLLER_IMAGE_TAG=${{ inputs.slim-controller-image-tag }} \
+          SLIM_CONTROLLER_CHART_TAG=${{ inputs.slim-controller-chart-tag }} \
           SLIM_CHART_TAG=${{ inputs.slim-chart-tag }} \
           KIND_CLUSTER_NAME=${{ inputs.kind-cluster-name }} \
           HELM_NAMESPACE=${{ inputs.kind-cluster-namespace }}
@@ -138,6 +143,7 @@ runs:
           SLIM_REPO_PATH='${{ inputs.slim-repo-path }}' \
           SLIM_IMAGE_TAG=${{ inputs.slim-image-tag }} \
           SLIM_CONTROLLER_IMAGE_TAG=${{ inputs.slim-controller-image-tag }} \
+          SLIM_CONTROLLER_CHART_TAG=${{ inputs.slim-controller-chart-tag }} \
           SLIM_CHART_TAG=${{ inputs.slim-chart-tag }} \
           KIND_CLUSTER_NAME=${{ inputs.kind-cluster-name }} \
           HELM_NAMESPACE=${{ inputs.kind-cluster-namespace }}
