+++ integrations/Taskfile.yml
@@ -18,13 +18,15 @@ includes:
     excludes: [ default ]
 
   slim:
-    taskfile: ./agntcy-slim/Taskfile.yml
-    dir: ./agntcy-slim
+    taskfile: ./agntcy-slim/topology/Taskfile.yml
+    dir: ./agntcy-slim/topology
     excludes: [ default ]
     vars:
       KIND_CLUSTER_NAME: '{{ .KIND_CLUSTER_NAME }}'
       SLIM_IMAGE_TAG: '{{ .SLIM_IMAGE_TAG }}'
       SLIM_CHART_TAG: '{{ .SLIM_CHART_TAG }}'
+      SLIM_CONTROLLER_IMAGE_TAG: '{{ .SLIM_CONTROLLER_IMAGE_TAG }}'
+      SLIM_CONTROLLER_CHART_TAG: '{{ .SLIM_CONTROLLER_CHART_TAG }}'
       HELM_NAMESPACE: '{{ .HELM_NAMESPACE }}'
       SLIM_REPO_PATH: '{{ .SLIM_REPO_PATH | default "" }}'
