+++ integrations/agntcy-slim/topology/config/peer-to-peer.yaml
