+++ integrations/agntcy-slim/topology/config/server-config.tpl
