+++ integrations/agntcy-slim/topology/examples/Dockerfile
